Add event bus
- Adapter 간 데이터 통신을 위한 Event bus 추가

+++ app/src/main/java/com/tedkim/smartschedule/model/ScheduleData.java
@@ -34,7 +34,7 @@ public class ScheduleData extends RealmObject {
 
     private int expectedDepartTime;
     private int totalTime;
-    private int distance;
+    private Double distance;
     private int beforeTime;
 
     public RealmList<RouteInfo> routeInfoList;
@@ -200,11 +200,11 @@ public class ScheduleData extends RealmObject {
         this.totalTime = totalTime;
     }
 
-    public int getDistance() {
+    public Double getDistance() {
         return distance;
     }
 
-    public void setDistance(int distance) {
+    public void setDistance(Double distance) {
         this.distance = distance;
     }
 }
